Spent a few hours or days figuring out how to rotate the rocket without rotating the camera. I originally had the camera as a child to the rocket, and figured I could get the camera to stop paying attention to the rocket's rotation via code, but I just couldn't get it. It was easier for me to un-child the camera, and use the script to modify it's position.
Discovered a bunch of things about coding in c# that is just too much to put into works completely. Methods do things, they live inside other methods or inside the level of code that initiates and runs. They tend to end with "();". Integers and floats are numbers, the former a whole and the latter a decimal. Time.deltatime has been used so far when I need to make something change over time, like when making the rocket go up or down. Mulitplying time.deltatime by a vector direction creates motion through space. This operation, written as Vector3.direction * time.deltatime, can be written within the parenthesis of a rigibody type, after choosing the rigibody type's property, such as "AddRelativeForce". It can also be written within the parenthesis of a transform type, after specifying the transform property (ie. Rotation, Position, etc). Sometimes, but not all, I was able to indicate an axis in which to work by adding a period to the end of the property and then the letter of the axis.

Most variables need to be declared at the beginning of the script, inside the class itself but outside of and before the first... uh... thing. If you want to use a variable to refer to a game object or property, you have to go through a long chain of referals. You have to "declare" the variable by typing the name of the type of variable you want to use then the name you want to use for the variable (using camel-case) at the top.

Then you have to define the variable inside of some automatically initiating process like "void Start" or "void Update". You do this by a couple means so far:

1) if it's a Unity component that's attached to the object the script is attached to you can type: variable = GetComponent<NameOfComponent>();

2)if it's a Game Object you can type: variable = GameObject.Find("Name of Game Object");       or           variable = GameObject.FindGameObjectWithTag("NameOfTag");

You can actually combine those two by writing it as variable = GameObject.Find("Name of Game Object").GetComponent<NameofComponent>();

Fucked around with the physics in the Unity inspector. It looks like a sweet spot, if I'm making this into some kind of jetpack descending/ascending down/up a hole game, is 1.5 mass and 1500 thrust power, which I may make the default.